# considerably speed up CPU matmul