Not sure how much we care about perf of the CPU version. This is trying to give a substantial boost, while still keeping complexity in check: without the comments, the  matmul still fits on a single screen.

While it would have been possible to either just print an error message for "bad" shapes, or write a function that is more generic and handles the unrollable part followed by an epilogue loop, I think the approach of having a slow, but very simple version, and a fast, but shape-restricted version is better. It keeps the fast algorithm more readable, and having the slow version there as a reference has pedagogical value in itself.

I have also removed the helper row pointers (`out_bt`), because they make it more difficult to see the strides of the individual memory accesses.

On my system, I observed about 20% end-to-end speedup (~5s to ~4s per step) with this change.