@azret 2024-05-14
@ngc92 I don't see the same performance improvements as you do.

```
== kernel #1 (nn.dev.MatMul_+matmul_forward_slow) ==

out:
OK: CPU: 51.56322 == GPU: 51.56322
OK: CPU: -13.7409 == GPU: -13.74089
OK: CPU: -2.222031 == GPU: -2.222029
OK: CPU: -13.50699 == GPU: -13.50698
OK: CPU: -18.66131 == GPU: -18.6613

OK

== kernel #2 (nn.dev.MatMul_+matmul_forward_new) ==

out:
OK: CPU: 51.56322 == GPU: 51.56321
OK: CPU: -13.7409 == GPU: -13.7409
OK: CPU: -2.222031 == GPU: -2.222035
OK: CPU: -13.50699 == GPU: -13.50699
OK: CPU: -18.66131 == GPU: -18.66132
ERROR: CPU: -67.68067 != GPU: -67.68078

FAILED.

kernel #1 forward (nn.dev.MatMul_+matmul_forward_slow), 3360.00 ms
kernel #2 forward (nn.dev.MatMul_+matmul_forward_new), 7265.00 ms
```

MSVC compiler does a pretty good job optimizing the matmul_forward_slow

--- Analyzing function: matmul_forward_slow
D:\SRC\nn.cs\nn\cpu\MatMul.c(15) : info C5001: loop vectorized
D:\SRC\nn.cs\nn\cpu\MatMul.c(13) : info C5002: loop not vectorized due to reason '1106'
D:\SRC\nn.cs\nn\cpu\MatMul.c(12) : info C5002: loop not vectorized due to reason '1106'

--- Analyzing function: matmul_forward
D:\SRC\nn.cs\nn\cpu\MatMul.c(46) : info C5002: loop not vectorized due to reason '1200'
D:\SRC\nn.cs\nn\cpu\MatMul.c(55) : info C5002: loop not vectorized due to reason '1200'
D:\SRC\nn.cs\nn\cpu\MatMul.c(53) : info C5002: loop not vectorized due to reason '1106'
D:\SRC\nn.cs\nn\cpu\MatMul.c(62) : info C5002: loop not vectorized due to reason '1200'
D:\SRC\nn.cs\nn\cpu\MatMul.c(43) : info C5002: loop not vectorized due to reason '1106'
D:\SRC\nn.cs\nn\cpu\MatMul.c(42) : info C5002: loop not vectorized due to reason '1106'