@ngc92 2024-05-14
This seems to indicate more that MSVC is doing a terrible job with the second loop, not a particularly good job at the first loop.  GCC produces the following assembly:
```asm
.L11:
vmovups ymm0, YMMWORD PTR [rcx+rax]
vfmadd231ps     ymm1, ymm0, YMMWORD PTR [rsi+rax]
vfmadd231ps     ymm2, ymm0, YMMWORD PTR [r8+rax]
vfmadd231ps     ymm3, ymm0, YMMWORD PTR [r10+rax]
vfmadd231ps     ymm4, ymm0, YMMWORD PTR [r14+rax]
vfmadd231ps     ymm5, ymm0, YMMWORD PTR [r13+0+rax]
vfmadd231ps     ymm6, ymm0, YMMWORD PTR [r12+rax]
vfmadd231ps     ymm7, ymm0, YMMWORD PTR [rbx+rax]
vfmadd231ps     ymm8, ymm0, YMMWORD PTR [r11+rax]
add     rax, 32
cmp     r9, rax
jne     .L11
```
This relies critically on `-Ofast', though, as it requires reordering floating-point summation. 

Could this be just because `msvc` is compiled without anything like `-match=native`, and thus runs in SSE2 mode without any AVX instructions? If I add `/arch:AVX2` as an option, I get
```asm
$LL13@matmul_for:
lea     eax, DWORD PTR [r14+r10]
movsxd  rcx, eax
movsxd  rax, r10d
vmovss  xmm0, DWORD PTR [r9+rcx*4]
vfmadd231ss xmm1, xmm0, DWORD PTR [rdi+rax*4]
lea     eax, DWORD PTR [r10+rdx]
movsxd  rcx, eax
lea     eax, DWORD PTR [r15+r10]
vfmadd231ss xmm2, xmm0, DWORD PTR [rdi+rcx*4]
movsxd  rcx, eax
lea     eax, DWORD PTR [r12+r10]
vfmadd231ss xmm3, xmm0, DWORD PTR [rdi+rcx*4]
movsxd  rcx, eax
lea     eax, DWORD PTR [r10+rbp]
vfmadd231ss xmm4, xmm0, DWORD PTR [rdi+rcx*4]
movsxd  rcx, eax
lea     eax, DWORD PTR [r10+r13]
vfmadd231ss xmm5, xmm0, DWORD PTR [rdi+rcx*4]
movsxd  rcx, eax
lea     eax, DWORD PTR [r11+r10]
vfmadd231ss xmm6, xmm0, DWORD PTR [rdi+rcx*4]
movsxd  rcx, eax
lea     eax, DWORD PTR [rbx+r10]
inc     r10d
vfmadd231ss xmm7, xmm0, DWORD PTR [rdi+rcx*4]
movsxd  rcx, eax
vfmadd231ss xmm8, xmm0, DWORD PTR [rdi+rcx*4]
sub     rsi, 1
jne     SHORT $LL13@matmul_for
```
not as nice as gccs code, but  still decent enough. With 4 issue slots, the leas and movs shouldn't actually slow down  the fmas. 
Here is a godbolt link to play around with this:
https://godbolt.org/z/vqhPqxaq4