@azret 2024-05-14
> This seems to indicate more that MSVC is doing a terrible job with the second loop, not a particularly good job at the first loop. GCC produces the following beautiful assembly:

👍