@rosslwheeler 2024-05-14
@ngc92 and @azret - you can add the /openmp:llvm flag so - /O2 /Oi /Ot /fp:fast /arch:AVX2 /openmp:llvm?