@azret 2024-05-14
@ngc92 Thanks for the link!

Yeah. MSVC does a poor job with your version :(

```
kernel #1 forward (nn.dev.MatMul_+matmul_forward_slow_avx_2), 2640.00 ms
kernel #2 forward (nn.dev.MatMul_+matmul_forward_ngc92_avx_2), 6938.00 ms
```