@dagelf 2024-05-21
It's still 3x slower than Pytorch... but it's a huge jump :smile: 

Proper way to run benchmarks:
```
( kill -STOP -1  # Stop all processes, NB don't run this outside a script!
timeout 40s ./train_gpt2
kill -CONT -1 )
```

Benchmark:
```
$ grep model /proc/cpuinfo |tail -1
model name	: Intel(R) Core(TM) i3-9100F CPU @ 3.60GHz

(this)
step 0: train loss 5.356185 (took 3973.912109 ms)
step 1: train loss 4.301033 (took 3499.634311 ms)
step 2: train loss 4.623316 (took 3498.866642 ms)
step 3: train loss 4.600415 (took 3541.733232 ms)
step 4: train loss 4.616777 (took 3501.966997 ms)
step 5: train loss 4.231482 (took 3500.156813 ms)

vs

(previous)
step 0: train loss 5.356185 (took 5332.631576 ms)
step 1: train loss 4.301033 (took 4840.134017 ms)
step 2: train loss 4.623316 (took 4828.423850 ms)
step 3: train loss 4.600415 (took 4828.398214 ms)
step 4: train loss 4.616777 (took 4829.080307 ms)
step 5: train loss 4.231482 (took 4858.988674 ms)
```