@dagelf 2024-05-21
@azret `$ grep model /proc/cpuinfo |tail -1` ?